- move value resolving to another class - refactoring ConfigurationManager - added missing javadoc to exist ones
